+++ packages/git-mob-core/src/git-mob-api/exec-command.ts
@@ -1,10 +1,19 @@
 import { exec } from 'node:child_process';
 import { promisify } from 'node:util';
+import { getConfig as cmGetConfig } from '../config-manager';
+
+type ExecCommandOptions = {
+  encoding: string;
+  cwd?: string;
+};
 
 // Runs the given command in a shell.
 export async function execCommand(command: string): Promise<string> {
+  const cmdConfig: ExecCommandOptions = { encoding: 'utf8' };
+  const processCwd = cmGetConfig('processCwd');
+  if (processCwd) cmdConfig.cwd = processCwd;
   const execAsync = promisify(exec);
-  const { stderr, stdout } = await execAsync(command, { encoding: 'utf8' });
+  const { stderr, stdout } = await execAsync(command, cmdConfig);
 
   if (stderr) {
     throw new Error(`GitMob execCommand: "${command}" ${stderr.trim()}`);